+++ Timereporter.Core/Models/WorkdayDto.cs
@@ -3,8 +3,8 @@
 	public class WorkdayDto
 	{
 		public string Date { get; set; }
-		public int ArrivalHours { get; set; }
-		public int BreakHours { get; set; }
-		public int DepartureHours { get; set; }
+		public float ArrivalHours { get; set; }
+		public float BreakHours { get; set; }
+		public float DepartureHours { get; set; }
 	}
 }
